requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
jsonpointer==2.1
mkl-service==2.4.2
protobuf==3.20.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
setuptools==69.5.1
wheel==0.43.0
requests
python-dotenv
Flask
gunicorn

